Added status handler to handler. renamed aHandler to handler.
