+++ src/components/P2P/handleProtocolCommands.ts
@@ -10,9 +10,8 @@ import { P2PCommandResponse } from '../../@types'
 import { P2P_CONSOLE_LOGGER } from './index.js'
 
 import { findDDO } from '../core/handlers/ddoHandler.js'
-import { NonceHandler, FeesHandler } from '../core/handlers/aHandler.js'
+import { NonceHandler, FeesHandler, StatusHandler } from '../core/handlers/handler.js'
 import { GENERIC_EMOJIS, LOG_LEVELS_STR } from '../../utils/logging/Logger.js'
-import { StatusHandler } from '../core/handlers/statusHandler.js'
 import { getConfig } from '../../utils/index.js'
 import { EncryptHandler } from '../core/handlers/encryptHandler.js'
 import { QueryHandler } from '../core/handlers/queryHandler.js'
